feat: Refactor financial models and services to use consistent naming and improve invoice calculations
