+++ src/main/java/br/com/hyteck/school_control/usecases/billing/GenerateConsolidatedStatementUseCase.java
@@ -1,15 +1,15 @@
 package br.com.hyteck.school_control.usecases.billing; // Ou pacote apropriado
 
+import br.com.hyteck.school_control.models.financial.Account;
+import br.com.hyteck.school_control.repositories.financial.LedgerEntryRepository;
+import br.com.hyteck.school_control.services.financial.AccountService;
 import br.com.hyteck.school_control.web.dtos.billing.ConsolidatedStatement;
 import br.com.hyteck.school_control.web.dtos.billing.StatementLineItem;
 import br.com.hyteck.school_control.models.payments.Invoice;
 import br.com.hyteck.school_control.models.payments.InvoiceStatus;
-import br.com.hyteck.school_control.models.finance.Account;
 import br.com.hyteck.school_control.models.payments.Responsible;
 import br.com.hyteck.school_control.repositories.InvoiceRepository;
-import br.com.hyteck.school_control.repositories.LedgerEntryRepository;
 import br.com.hyteck.school_control.repositories.ResponsibleRepository;
-import br.com.hyteck.school_control.services.AccountService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
