+++ src/main/java/br/com/hyteck/school_control/web/dtos/billing/InvoiceStatusSummaryDto.java
@@ -20,26 +20,6 @@ record ResponsibleInfo(String id, String name) {
 }
 
 /**
- * Represents basic information about an invoice, including its current ledger balance.
- *
- * @param id                The unique identifier of the invoice.
- * @param responsibleId     The ID of the responsible party.
- * @param responsibleName   The name of the responsible party.
- * @param currentBalanceDue The current balance due on the invoice, derived from ledger entries.
- * @param dueDate           The due date of the invoice.
- * @param status            The current status of the invoice.
- */
-record InvoiceBasicInfo(
-        String id,
-        String responsibleId,
-        String responsibleName,
-        BigDecimal currentBalanceDue,
-        LocalDate dueDate,
-        InvoiceStatus status
-) {
-}
-
-/**
  * Data Transfer Object for providing a summary of invoice statuses.
  * This DTO can hold various counts, total balances, and lists of invoices
  * based on their status (Pending, Overdue, Paid).
